Fix a couple weird return value types

+++ common/m_fileio.h
@@ -36,7 +36,7 @@ std::string M_FindUserFileName(const std::string& file, const char* ext);
 void M_FixPathSep(std::string& path);
 std::string M_GetCWD();
 
-int32_t M_FileLength (FILE *f);
+int64_t M_FileLength (FILE *f);
 bool M_FileExists(const std::string& filename);
 bool M_FileExistsExt(const std::string& filename, const char* ext);
